+++ server/graphql/repack/mutations/insert.rs
@@ -16,7 +16,7 @@ use service::{
 pub struct InsertRepackInput {
     pub stock_line_id: String,
     pub number_of_packs: f64,
-    pub new_pack_size: i32,
+    pub new_pack_size: f64,
     pub new_location_id: Option<String>,
 }
