+++ server/service/src/invoice_line/stock_in_line/insert/mod.rs
@@ -24,7 +24,7 @@ pub struct InsertStockInLine {
     pub invoice_id: String,
     pub item_id: String,
     pub location: Option<NullableUpdate<String>>,
-    pub pack_size: u32,
+    pub pack_size: f64,
     pub batch: Option<String>,
     pub note: Option<String>,
     pub cost_price_per_pack: f64,
@@ -180,7 +180,7 @@ mod test {
                 &context,
                 inline_init(|r: &mut InsertStockInLine| {
                     r.id = "new invoice line id".to_string();
-                    r.pack_size = 0;
+                    r.pack_size = 0.0;
                 }),
             ),
             Err(ServiceError::PackSizeBelowOne)
@@ -192,7 +192,7 @@ mod test {
                 &context,
                 inline_init(|r: &mut InsertStockInLine| {
                     r.id = "new invoice line id".to_string();
-                    r.pack_size = 1;
+                    r.pack_size = 1.0;
                     r.number_of_packs = 0.0;
                 }),
             ),
@@ -205,7 +205,7 @@ mod test {
                 &context,
                 inline_init(|r: &mut InsertStockInLine| {
                     r.id = "new invoice line id".to_string();
-                    r.pack_size = 1;
+                    r.pack_size = 1.0;
                     r.number_of_packs = 1.0;
                     r.item_id = "invalid".to_string();
                 }),
@@ -219,7 +219,7 @@ mod test {
                 &context,
                 inline_init(|r: &mut InsertStockInLine| {
                     r.id = "new invoice line id".to_string();
-                    r.pack_size = 1;
+                    r.pack_size = 1.0;
                     r.number_of_packs = 1.0;
                     r.item_id = mock_item_a().id;
                     r.location = Some(NullableUpdate {
@@ -236,7 +236,7 @@ mod test {
                 &context,
                 inline_init(|r: &mut InsertStockInLine| {
                     r.id = "new invoice line id".to_string();
-                    r.pack_size = 1;
+                    r.pack_size = 1.0;
                     r.number_of_packs = 1.0;
                     r.item_id = mock_item_a().id;
                     r.invoice_id = "new invoice id".to_string();
@@ -252,7 +252,7 @@ mod test {
                 inline_init(|r: &mut InsertStockInLine| {
                     r.id = "new invoice line id".to_string();
                     r.item_id = mock_item_a().id.clone();
-                    r.pack_size = 1;
+                    r.pack_size = 1.0;
                     r.number_of_packs = 1.0;
                     r.invoice_id = mock_outbound_shipment_e().id;
                 }),
@@ -267,7 +267,7 @@ mod test {
                 inline_init(|r: &mut InsertStockInLine| {
                     r.id = "new invoice line id".to_string();
                     r.item_id = mock_item_a().id.clone();
-                    r.pack_size = 1;
+                    r.pack_size = 1.0;
                     r.number_of_packs = 1.0;
                     r.invoice_id = verified_inbound_return().id; // VERIFIED
                 }),
@@ -281,7 +281,7 @@ mod test {
                 &context,
                 inline_init(|r: &mut InsertStockInLine| {
                     r.id = "new invoice line id".to_string();
-                    r.pack_size = 1;
+                    r.pack_size = 1.0;
                     r.number_of_packs = 1.0;
                     r.item_id = mock_item_a().id;
                     r.invoice_id = mock_inbound_return_a().id; // Store B
@@ -309,7 +309,7 @@ mod test {
                 r.id = "new_invoice_line_id".to_string();
                 r.invoice_id = mock_inbound_return_a().id;
                 r.item_id = mock_item_a().id;
-                r.pack_size = 1;
+                r.pack_size = 1.0;
                 r.number_of_packs = 1.0;
                 r.barcode = Some(gtin.clone());
             }),
@@ -333,7 +333,7 @@ mod test {
             inline_edit(&inbound_line, |mut u| {
                 u.id = "new_invoice_line_id".to_string();
                 u.item_link_id = mock_item_a().id;
-                u.pack_size = 1;
+                u.pack_size = 1.0;
                 u.number_of_packs = 1.0;
                 u
             })
@@ -364,7 +364,7 @@ mod test {
                 r.id = "new_invoice_line_pack_to_one".to_string();
                 r.invoice_id = mock_inbound_return_a().id;
                 r.item_id = mock_item_a().id;
-                r.pack_size = 10;
+                r.pack_size = 10.0;
                 r.number_of_packs = 20.0;
                 r.sell_price_per_pack = 100.0;
             }),
@@ -380,7 +380,7 @@ mod test {
             inline_edit(&inbound_line, |mut u| {
                 u.id = "new_invoice_line_pack_to_one".to_string();
                 u.item_link_id = mock_item_a().id;
-                u.pack_size = 1;
+                u.pack_size = 1.0;
                 u.number_of_packs = 200.0;
                 u.sell_price_per_pack = 10.0;
                 u
